Fix for build with -qreal float
Resolve ambiguity with function calls (due to double->float conversion),
use qreal where needed intead of explicit 'double'.

Task-number: QTBUG-47433

+++ src/widgets/widgets/qlineedit_p.cpp
@@ -343,7 +343,7 @@ void QLineEditIconButton::startOpacityAnimation(qreal endValue)
 void QLineEditIconButton::updateCursor()
 {
 #ifndef QT_NO_CURSOR
-    setCursor(qFuzzyCompare(m_opacity, 1.0) || !parentWidget() ? QCursor(Qt::ArrowCursor) : parentWidget()->cursor());
+    setCursor(qFuzzyCompare(m_opacity, qreal(1.0)) || !parentWidget() ? QCursor(Qt::ArrowCursor) : parentWidget()->cursor());
 #endif
 }
